requirements.txt
aniso8601==6.0.0
asn1crypto==0.24.0
attrs==19.1.0
aws-xray-sdk==0.95
awsebcli==3.15.0
backcall==0.1.0
blessed==1.15.0
boto==2.49.0
boto3==1.9.134
botocore==1.12.134
cached-property==1.5.1
cement==2.8.2
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.3.9
cryptography==2.6.1
decorator==4.4.0
docker==3.7.2
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
ecdsa==0.13.2
envs==1.3
Flask==1.0.2
Flask-API==1.1
flask-restplus==0.12.1
future==0.16.0
idna==2.7
ipython==7.4.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
jsondiff==1.1.1
jsonpickle==1.1
jsonschema==2.6.0
MarkupSafe==1.1.1
mock==3.0.5
moto==1.3.7
parso==0.4.0
pathspec==0.5.9
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyaml==19.4.1
pycparser==2.19
pycryptodome==3.3.1
Pygments==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
python-dotenv==0.10.1
python-jose==2.0.2
python-jose-cryptodome==1.3.2
pytz==2019.1
PyYAML==3.13
requests==2.20.1
responses==0.10.6
rope==0.14.0
s3transfer==0.2.0
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
texttable==0.9.1
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.24.2
warrant==0.6.1
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.2
wrapt==1.11.1
xmltodict==0.12.0
